refactor(schemes): zoo package + Machines split + concurrency proof (PR review)
Addresses every inline thread on PR #24:

- One entity per file in the new schemes.zoo package: Attr, Coattr,
  Gather, Scatter, Cata, Ana, FoldM, CataM, AnaM (Citizens/CitizensM/
  Decorations.scala dissolved). FoldM is now PUBLIC with an open
  constructor — users can wrap their own S => M[A] as a citizen;
  CataM/AnaM still extend it (no longer sealed, noted in scaladoc).

- Schemes.scala keeps only the factory methods; ALL machinery moved to
  private[schemes] object Machines (OnStackLimit, EmptyAnyRefs,
  childrenArr, AscendToken, rebuildLayer(+Paired), foldLayered/Or/M,
  fusedPairedFold(+M)). Machines' header documents the thread-safety
  model: every machine allocates its mutable state per invocation (per
  FORCE on the M path), the only shared values are immutable sentinels.
  EmptyAnyRefs' scaladoc answers what/who/why-threadsafe directly (the
  PR question): a zero-length leaf sentinel used by childrenArr in every
  engine; read-only by construction (store loops bound at arr.length).

- The thread-safety claim is now TESTED, both places the review asked:
  SchemesConcurrencySpec (16 concurrent tasks mixing cata/hylo/ana/
  para/histo/futu/cataM[Eval] over shared optics and fixtures) and
  PlatedConcurrencySpec in tests/ (concurrent Plated transform/universe/
  rewrite over a shared structure).

Fixed in review of the restructure itself: the concurrency spec's
coalgebra guarded at n <= 0, making seed 1 its own child — a
non-terminating unfold the engines correctly turn into heap exhaustion.
Guard is n <= 1; the suite runs in the default 1GB test heap.

All 511 tests green; mdoc clean.
